fix: Fix and refactor GranularHttpSessionServlet and remove the redundant abstract common class.
